==> 02-Embeddings.ipynb <==
from __future__ import absolute_import, division, print_function, unicode_literals

import os
import sys
import pickle

import numpy as np
import pandas as pd

from sklearn.model_selection import train_test_split
from sklearn.pipeline import Pipeline
from sklearn.base import BaseEstimator, TransformerMixin
from sklearn.metrics import accuracy_score, recall_score, \
                                precision_score, f1_score

import tensorflow as tf
from tensorflow import keras
from tensorflow.keras import layers

from tensorflow.keras.preprocessing.text import Tokenizer
from tensorflow.keras.preprocessing.sequence import pad_sequences

import tensorflow_datasets as tfds
tfds.disable_progress_bar()

import warnings
warnings.filterwarnings('ignore')

### Haven`t find better way to download encoder

(train_data, test_data), info = tfds.load(
    'imdb_reviews/subwords8k',
    split=(tfds.Split.TRAIN, tfds.Split.TEST),
    with_info=True, as_supervised=True
)

encoder = info.features['text'].encoder

with open(os.path.join('pickles', 'encoder.pickle'), 'wb') as handle:
    pickle.dump(encoder, handle, protocol=pickle.HIGHEST_PROTOCOL)

[f for f in os.listdir('DATA') if not f.startswith('.')]

def fetch_data_train(path='DATA'):
    data = pd.read_csv(os.path.join(path, 'train.csv'))
    data_positive = data.loc[data['target'] == 1][:80000]
    data_negative = data.loc[data['target'] == 0][:80000]
    
    data = data_positive.append(data_negative, ignore_index=True, sort=False)
    data = data.sample(frac=1).reset_index(drop=True)
    
    X = data.drop(['qid', 'target'], axis=1).values
    y = data['target'].values
    
    X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2)
    
    return X_train, X_test, y_train, y_test

def fetch_data_test(path='DATA'):
    train = pd.read_csv(os.path.join(path, 'train.csv'))
    test = pd.read_csv(os.path.join(path, 'test.csv'))
    
    X_train = train.drop(['qid', 'target'], axis=1).values
    y_train = train['target'].values
    X_test = train.drop(['qid'], axis=1).values
    
    return X_train, X_test, y_train

X_train, X_test, y_train, y_test = fetch_data_train()

X_train

example_sentence = X_train[0][0]
encoded_sentence = encoder.encode(example_sentence)
decoded_sentence = encoder.decode(encoded_sentence)

assert example_sentence == decoded_sentence

print('Original sentence:', decoded_sentence)
print('Encoded sentence', encoded_sentence)

## Custom embedding pipeline

class EmbeddingTransformer(BaseEstimator, TransformerMixin):
    def __init__(self, encoder):
        self.encoder = encoder
        self.vocab_size = self.encoder.vocab_size
    
    def fit(self, X, y=None):
        return self
    
    def transform(self, X, y=None):
        X_transformed = list()
        for line in X.reshape(1, -1)[0]:
            X_transformed.append(
                self.encoder.encode(line)
            )
        return np.array(X_transformed)

class PaddingTransform(BaseEstimator, TransformerMixin):
    def __init__(self, max_len=None):
        if max_len is not None:
            self.max_len = max_len
    
    def fit(self, X, y=None):
        if 'max_len' not in self.__dir__():
            self.max_len = len(max(X, key=len))
        return self
    
    def transform(self, X, y=None):
        X_transformed = []
        
        for line in X:
            X_transformed.append(
                np.concatenate((
                    np.array(line),
                    np.array([0] * (self.max_len - len(line)))
                ), axis=0)
            )
        return np.array(X_transformed, dtype=np.int32)

pipeline = Pipeline([
    ('embedding', EmbeddingTransformer(encoder=encoder)),
    ('padding', PaddingTransform())
])

pipeline.fit(np.concatenate((
    X_train, X_test
), axis=0))
X_train = pipeline.transform(X_train)
X_test = pipeline.transform(X_test)

# Use part of data for validation
X_valid = X_train[-1000:]
y_valid = y_train[-1000:]
X_train = X_train[:-1000]
y_train = y_train[:-1000]

print('X shape:', X_train.shape)
X_train

to_save = {
    'X_train': X_train,
    'y_train': y_train,
    'X_valid': X_valid,
    'y_valid': y_valid,
    'X_test': X_test,
    'y_test': y_test,
}

with open(os.path.join('pickles', 'dataset.pickle'), 'wb') as handle:
    pickle.dump(to_save, handle, protocol=pickle.HIGHEST_PROTOCOL)

## Some popular solutions

#### Basic preprocessing

X_train, X_test, y_train, y_test = fetch_data_train()

embed_size = 300
max_features = 50000
maxlen = 100

tokenizer = Tokenizer(num_words=max_features)

tokenizer.fit_on_texts(X_train.reshape(1, -1)[0])

X_train = tokenizer.texts_to_sequences(X_train.reshape(1, -1)[0])
X_test = tokenizer.texts_to_sequences(X_test.reshape(1, -1)[0])

X_train = pad_sequences(X_train, maxlen=maxlen)
X_test = pad_sequences(X_test, maxlen=maxlen)

X_train

### Stanford`s Glove

file = os.path.join(*['DATA', 'embeddings', 'glove.840B.300d', 'glove.840B.300d.txt'])
def load_coefs(word, *vect):
    return word, np.array(vect, dtype=np.float32)
embeddings = dict(load_coefs(*w.split(" ")) for w in open(file, 'r'))

print('Number of embeddings:', len(embeddings.keys()))
print('Embedding vector shape:', embeddings['foo'].shape)

all_embs = np.stack(embeddings.values())
emb_mean, emb_std = all_embs.mean(), all_embs.std()
embed_size = all_embs.shape[1]

all_embs

word_index = tokenizer.word_index
nb_words = min(max_features, len(word_index))

embedding_matrix = np.random.normal(emb_mean, emb_std, (nb_words, embed_size))

for word, i in word_index.items():
    if i >= max_features:
        continue
    embedding_vector = embeddings_index.get(word)
    if embedding_vector is not None:
        embedding_matrix[i] = embedding_vector
        
embedding_matrix = embedding_matrix.astype(np.float32)

embedding_matrix

embedding_dim = 16

model = keras.Sequential([
    layers.Embedding(nb_words, embed_size),
    layers.GlobalAveragePooling1D(),
    layers.Dense(16, activation='relu'),
    layers.Dense(1, activation='sigmoid')
])

model.summary()

model.compile(optimizer='adam',
              loss='binary_crossentropy',
              metrics=['accuracy'])

history = model.fit(
    X_train, y_train,
    epochs=1,
    batch_size=32,
)

y_pred = model.predict(X_test)
y_pred[np.where(y_pred >= 0.5)] = 1
y_pred[np.where(y_pred < 0.5)] = 0

print('Accuracy:', accuracy_score(y_test, y_pred))
print('Recall:', recall_score(y_test, y_pred))
print('Precision:', precision_score(y_test, y_pred))
print('F1 Score:', f1_score(y_test, y_pred))
