==> 03-Building-Model.ipynb <==
from __future__ import absolute_import, division, print_function, unicode_literals

import os
import sys
import pickle

import numpy as np
import pandas as pd

from sklearn.preprocessing import MinMaxScaler
from sklearn.metrics import accuracy_score, recall_score, \
                                precision_score, f1_score

import tensorflow as tf
from tensorflow import keras
from tensorflow.keras import layers

import matplotlib.pyplot as plt
%matplotlib inline

### Loading preprocessed Data

with open(os.path.join('pickles', 'dataset.pickle'), 'rb') as handle:
    dataset = pickle.load(handle)

X_train = dataset['X_train']
y_train = dataset['y_train']
X_valid = dataset['X_valid']
y_valid = dataset['y_valid']
X_test = dataset['X_test']
y_test = dataset['y_test']

with open(os.path.join('pickles', 'encoder.pickle'), 'rb') as handle:
    encoder = pickle.load(handle)

## Building Model 

embedding_dim = 16

model = keras.Sequential([
    layers.Embedding(encoder.vocab_size, embedding_dim),
    layers.GlobalAveragePooling1D(),
    layers.Dense(16, activation='relu'),
    layers.Dense(1, activation='sigmoid')
])

model.summary()

model.compile(optimizer='adam',
              loss='binary_crossentropy',
              metrics=['accuracy'])

early_stop = keras.callbacks.EarlyStopping(monitor='val_loss', patience=10)

history = model.fit(
    X_train, y_train,
    epochs=10,
    batch_size=32,
    validation_data=(X_valid, y_valid),
    callbacks=[early_stop]
)

y_pred = model.predict(X_test)

y_pred = y_pred.reshape(1, -1)[0]
y_pred[np.where(y_pred >= 0.5)] = 1
y_pred[np.where(y_pred < 0.5)] = 0

print('Accuracy:', accuracy_score(y_test, y_pred))
print('Recall:', recall_score(y_test, y_pred))
print('Precision:', precision_score(y_test, y_pred))
print('F1 Score:', f1_score(y_test, y_pred))

weights = model.layers[0]\
                .get_weights()[0]
vect_file = os.path.join('embeddings', 'vecs.tsv')
meta_file = os.path.join('embeddings', 'meta.tsv')

with open(vect_file, 'w') as out_vect, open(meta_file, 'w') as out_meta:
    for i, word in enumerate(encoder.subwords):
        vec = weights[num + 1]
        out_meta.write(word + '\n')
        out_vect.write('\t'.join([str(w) for w in vec]) + '\n')
